fix: post issue

+++ src/main/java/com/cams/todo/services/impl/TodoServiceImpl.java
@@ -7,15 +7,17 @@ import com.cams.todo.entities.Todo;
 import com.cams.todo.exceptions.TodoNotFoundException;
 import com.cams.todo.mappers.TodoMapper;
 import com.cams.todo.repositories.TodoRepository;
+import org.springframework.transaction.annotation.Transactional;
 import com.cams.todo.services.TodoService;
 import lombok.AllArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
+
 @Service
 @AllArgsConstructor
+@Transactional
 public class TodoServiceImpl implements TodoService {
 
     private final TodoMapper todoMapper;
@@ -29,9 +31,9 @@ public class TodoServiceImpl implements TodoService {
     }
 
     @Override
+    @Transactional(readOnly = true)
     public Page<TodoResponse> getTodos(Pageable pageable) {
-        return todoRepository
-                .findAll(pageable)
+        return todoRepository.findAll(pageable)
                 .map(todoMapper::toResponse);
     }
 
@@ -41,13 +43,16 @@ public class TodoServiceImpl implements TodoService {
                 .orElseThrow(() -> new TodoNotFoundException(id));
 
         todoMapper.updateEntityFromDto(dto, entity);
-
         Todo saved = todoRepository.save(entity);
+
         return todoMapper.toResponse(saved);
     }
 
     @Override
     public void deleteById(Long id) {
+        if (!todoRepository.existsById(id)) {
+            throw new TodoNotFoundException(id);
+        }
         todoRepository.deleteById(id);
     }
 }
